prints

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -6,7 +6,7 @@
 */
 int main(void)
 {
-int n, fp;
+long int n, fp;
 n = 612852475143;
 for (fp = 2; fp <= n; fp++)
 {
@@ -16,6 +16,6 @@ n /= fp;
 fp--;
 }
 }
-printf("%1d\n", fp);
+printf("%ld\n", fp);
 return (0);
 }
